move class - ProfileImageFile
