+++ src/main/java/co/fineants/api/infra/s3/service/imple/AmazonS3WriteProfileImageFileService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import co.fineants.api.domain.holding.domain.factory.UuidGenerator;
+import co.fineants.api.domain.member.domain.entity.ProfileImageFile;
 import co.fineants.api.infra.s3.service.RemoteFileUploader;
 import co.fineants.api.infra.s3.service.WriteProfileImageFileService;
