+++ src/test/java/co/fineants/api/infra/s3/service/imple/AmazonS3WriteProfileImageFileServiceTest.java
@@ -11,6 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import co.fineants.TestDataFactory;
 import co.fineants.api.domain.holding.domain.factory.UuidGenerator;
+import co.fineants.api.domain.member.domain.entity.ProfileImageFile;
 import co.fineants.api.global.errors.exception.business.ImageEmptyInvalidInputException;
 import co.fineants.api.infra.s3.service.RemoteFileUploader;
 import co.fineants.api.infra.s3.service.WriteProfileImageFileService;
